requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
crawl4ai>=0.3.0
gunicorn==21.2.0
python-dotenv==1.0.0
psutil==5.9.6